Continued trying to isolate points based on whether or not they are in cosmos/zcosmos/both.
I have two slightly different methods of going about this. If I can get them to work, the first will leave me with a dataframe including the data for points in both cosmos and zcosmos, and the second method will produce a document including the data for points ONLY in cosmos or ONLY zcosmos.