Finished fixing the crossmatching and isolating data based on whether it was in cosmos/zcosmos/both, as well as did a little housekeeping.
I made a dataframe with all data for points in cosmos also in zcosmos and had the zcosmos and cosmos numbers alighned. I made a dataframe for all points ONLY in cosmos and for all points ONLY in zcosmos.
I also did some cleaning of code I had, moving uneccessary code to my extra_code file and making all names in lowercase because it turns out Jupyter is super obessed with capitalization.